All tests passing
Need to finish running on targets, but make tests is working locally.
Had to #ifdef some platform-specific code that isn't in /platforms
(yet).

+++ src/interface/interface.cpp
@@ -13,7 +13,7 @@ static const char interfaceNames[][5] = {
 };
 
 const char* openxc::interface::descriptorToString(InterfaceDescriptor* descriptor) {
-    if(descriptor->type < sizeof(interfaceNames) / sizeof(interfaceNames[0])) {
+    if((int)descriptor->type < sizeof(interfaceNames) / sizeof(interfaceNames[0])) {
         return interfaceNames[descriptor->type];
     }
     return "Unknown";
